+++ src/test/kotlin/html4tree/UtilTest.kt
@@ -1,130 +1,98 @@
 package html4tree
 
 import org.junit.Test
+import java.io.File
 import kotlin.test.assertEquals
 import kotlin.test.assertNull
-import java.io.File
 
 class UtilTest {
 
     @Test
     fun testLinkedList() {
-        val ll = LinkedList()
+        val list = LinkedList()
+        assertNull(list.pull())
 
-        // 초기에 pull() 하면 null이 나와야 함
-        assertNull(ll.pull())
+        val file1 = File("file1")
+        val file2 = File("file2")
+        val file3 = File("file3")
 
-        val file1 = File("test1")
-        val file2 = File("test2")
+        list.push(LinkedListEntry(file1, 0))
+        list.push(LinkedListEntry(file2, 1))
+        list.push(LinkedListEntry(file3, 2))
 
+        var pulled = list.pull()
+        assertEquals(file1, pulled?.file)
+        assertEquals(0, pulled?.level)
+
+        pulled = list.pull()
+        assertEquals(file2, pulled?.file)
+        assertEquals(1, pulled?.level)
+
+        pulled = list.pull()
+        assertEquals(file3, pulled?.file)
+        assertEquals(2, pulled?.level)
+        assertNull(list.pull())
+    }
+
+    @Test
+    fun testEntryDataClass() {
+        val file1 = File("file1")
+        val entry1 = Entry(file1, 0, null)
+        val entry2 = Entry(file1, 0, null)
+
+        assertEquals(entry1, entry2)
+        assertEquals("Entry(data=file1, level=0, next=null)", entry1.toString())
+    }
+
+    @Test
+    fun testLinkedListEntryDataClass() {
+        val file1 = File("file1")
         val entry1 = LinkedListEntry(file1, 0)
-        val entry2 = LinkedListEntry(file2, 1)
+        val entry2 = LinkedListEntry(file1, 0)
 
-        ll.push(entry1)
-        ll.push(entry2)
-
-        // Queue처럼 처음 들어간 것부터 나와야 함
-        val out1 = ll.pull()
-        assertEquals(file1, out1?.file)
-        assertEquals(0, out1?.level)
-
-        val out2 = ll.pull()
-        assertEquals(file2, out2?.file)
-        assertEquals(1, out2?.level)
-
-        assertNull(ll.pull())
+        assertEquals(entry1, entry2)
+        assertEquals("LinkedListEntry(file=file1, level=0)", entry1.toString())
     }
 
     @Test
-    fun testEscapeHtml() {
-        assertEquals("&amp;", "&".escapeHtml())
-        assertEquals("&lt;", "<".escapeHtml())
-        assertEquals("&gt;", ">".escapeHtml())
-        assertEquals("&quot;", "\"".escapeHtml())
-        assertEquals("&#x27;", "'".escapeHtml())
-        assertEquals("&lt;div class=&quot;test&quot;&gt;Hello &amp; Welcome&#x27;&lt;/div&gt;", "<div class=\"test\">Hello & Welcome'</div>".escapeHtml())
+    fun testLinkedListPushExisting() {
+        val list = LinkedList()
+        list.push(LinkedListEntry(File("f1"), 0))
+        list.push(LinkedListEntry(File("f2"), 0))
+        val entry1 = list.pull()
+        val entry2 = list.pull()
+        assertEquals(File("f1"), entry1?.file)
+        assertEquals(File("f2"), entry2?.file)
     }
 
     @Test
-    fun testUrlEncodePath() {
-        assertEquals("hello%20world", "hello world".urlEncodePath())
-        assertEquals("test%2B%3F%26", "test+?&".urlEncodePath())
+    fun testLinkedListAccessors() {
+        val list = LinkedList()
+        list.first = Entry(File("test"), 0, null)
+        list.last = Entry(File("test"), 0, null)
+        assertEquals(File("test"), list.first?.data)
+        assertEquals(File("test"), list.last?.data)
     }
-
-
-    @Test
-    fun testLinkedListFull() {
-        val ll = LinkedList()
-        val file1 = File("f1")
-        val entry1 = LinkedListEntry(file1, 0)
-        ll.push(entry1)
-        val file2 = File("f2")
-        val entry2 = LinkedListEntry(file2, 1)
-        ll.push(entry2)
-        val file3 = File("f3")
-        val entry3 = LinkedListEntry(file3, 2)
-        ll.push(entry3)
-
-        assertEquals(file1, ll.pull()?.file)
-        assertEquals(file2, ll.pull()?.file)
-        assertEquals(file3, ll.pull()?.file)
-        assertNull(ll.pull())
-    }
-
-
-    @Test
-    fun testLinkedListPushFirstNotNull() {
-        val ll = LinkedList()
-        val entry1 = LinkedListEntry(File("f1"), 0)
-        ll.push(entry1)
-        val entry2 = LinkedListEntry(File("f2"), 0)
-        ll.push(entry2)
-        assertEquals(File("f1"), ll.pull()?.file)
-        assertEquals(File("f2"), ll.pull()?.file)
-    }
-
-    @Test
-    fun testLinkedListPullNextNotNull() {
-        val ll = LinkedList()
-        val entry1 = LinkedListEntry(File("f1"), 0)
-        ll.push(entry1)
-        val entry2 = LinkedListEntry(File("f2"), 0)
-        ll.push(entry2)
-        ll.pull()
-        ll.pull()
-        assertNull(ll.pull())
-    }
-
-
 
     @Test
     fun testLinkedListPushNullFirst() {
-        val ll = LinkedList()
-        ll.first = null
-        ll.last = Entry(File("tmp"), 0, null)
-        val entry = LinkedListEntry(File("tmp2"), 0)
-        ll.push(entry)
-        assertNull(ll.first?.next)
+        val list = LinkedList()
+        list.last = Entry(File("fake"), 0, null)
+        list.push(LinkedListEntry(File("f3"), 0))
+        assertEquals(File("fake"), list.pull()?.file)
+        assertEquals(File("f3"), list.pull()?.file)
+        assertEquals(File("f3"), list.first?.data)
     }
 
     @Test
-    fun testLinkedListPullNullFirst() {
-        val ll = LinkedList()
-        ll.first = null
-        ll.last = Entry(File("tmp"), 0, null)
-        val pulled = ll.pull()
-        assertEquals(File("tmp"), pulled?.file)
+    fun testLinkedListPushNullFirstWithExistingChain() {
+        val list = LinkedList()
+        list.last = Entry(File("f1"), 0, Entry(File("f2"), 0, null))
+        list.push(LinkedListEntry(File("f3"), 0))
+
+        assertEquals(File("f1"), list.pull()?.file)
+        assertEquals(File("f2"), list.pull()?.file)
+        assertEquals(File("f3"), list.pull()?.file)
+        assertNull(list.pull())
     }
-
-
-
-    @Test
-    fun testLinkedListEntryConstructors() {
-        val ll = LinkedList()
-        val e = Entry(File("a"), 0, null)
-        ll.first = e
-        ll.last = e
-        assertEquals(e, ll.first)
-    }
-
 }